+++ app/pages/preview/contextCreator.js
@@ -1,6 +1,6 @@
 import manifest from './manifest.json';
 import { baseUrl } from '../../config';
-import { formatDate } from '../../helpers/dateFormatter';
+import { formatDate } from '../../helpers/common/dateFormatter';
 import { formatPrice } from '../../helpers/contextCreators/priceFormatter';
 import { logger } from '../../logger';
