+++ app/pages/preview/ui-tests/general.ui.test.js
@@ -3,7 +3,7 @@ import { ClientFunction, Selector } from 'testcafe';
 import { extractInnerText } from 'buying-catalogue-library';
 import content from '../manifest.json';
 import { orderApiUrl } from '../../../config';
-import { formatDate } from '../../../helpers/dateFormatter';
+import { formatDate } from '../../../helpers/common/dateFormatter';
 
 const pageUrl = 'http://localhost:1234/order/organisation/order-1/preview';
